Move more methods from EthereumManager to EvmManager
* Removed 'ethereum'/'eth' from f-strings, comments and method names
  such as `get_historical_eth_balance` `query_eth_highest_block`,
  `get_multieth_balance`, `get_eth_balance`
