+++ rotkehlchen/chain/evm/contracts.py
@@ -20,9 +20,9 @@ from rotkehlchen.chain.ethereum.abi import decode_event_data_abi
 from rotkehlchen.types import ChecksumEvmAddress
 
 if TYPE_CHECKING:
-    from rotkehlchen.chain.ethereum.manager import EthereumManager
     from rotkehlchen.chain.ethereum.structures import EthereumTxReceiptLog
     from rotkehlchen.chain.ethereum.types import WeightedNode
+    from rotkehlchen.chain.evm.manager import EvmManager
 
 WEB3 = Web3()
 
@@ -34,7 +34,7 @@ class EvmContract(NamedTuple):
 
     def call(
             self,
-            manager: 'EthereumManager',
+            manager: 'EvmManager',
             method_name: str,
             arguments: Optional[List[Any]] = None,
             call_order: Optional[Sequence['WeightedNode']] = None,
@@ -51,7 +51,7 @@ class EvmContract(NamedTuple):
 
     def get_logs(
             self,
-            manager: 'EthereumManager',
+            manager: 'EvmManager',
             event_name: str,
             argument_filters: Dict[str, Any],
             from_block: int,
